Issue #24562 move MappedSuperclass entities and tests to Data JPA tests
