+++ dev/io.openliberty.data.internal_fat_jpa/test-applications/DataJPATestApp/src/test/jakarta/data/jpa/web/DataJPATestServlet.java
@@ -22,6 +22,7 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -73,6 +74,9 @@ public class DataJPATestServlet extends FATServlet {
 
     @Inject
     ShippingAddresses shippingAddresses;
+
+    @Inject
+    Tariffs tariffs;
 
     @Resource
     private UserTransaction tran;
@@ -1066,6 +1070,186 @@ public class DataJPATestServlet extends FATServlet {
         assertIterableEquals(List.of("Monica@tests.openliberty.io",
                                      "martin@tests.openliberty.io"),
                              creditCards.findByExpiresOnBetween(LocalDate.of(2026, 1, 1), LocalDate.of(2026, 12, 31)));
+    }
+
+    /**
+     * Add, find, and remove entities with a mapped superclass.
+     * Also tests automatically paginated iterator and list.
+     */
+    @Test
+    public void testMappedSuperclass() {
+        tariffs.deleteByLeviedBy("USA");
+
+        Tariff t1 = new Tariff();
+        t1.leviedAgainst = "China";
+        t1.leviedBy = "USA";
+        t1.leviedOn = "Solar Panels";
+        t1.rate = 0.15f;
+        tariffs.save(t1);
+
+        Tariff t2 = new Tariff();
+        t2.leviedAgainst = "Germany";
+        t2.leviedBy = "USA";
+        t2.leviedOn = "Steel";
+        t2.rate = 0.25f;
+        tariffs.save(t2);
+
+        Tariff t3 = new Tariff();
+        t3.leviedAgainst = "India";
+        t3.leviedBy = "USA";
+        t3.leviedOn = "Aluminum";
+        t3.rate = 0.1f;
+        tariffs.save(t3);
+
+        Tariff t4 = new Tariff();
+        t4.leviedAgainst = "Japan";
+        t4.leviedBy = "USA";
+        t4.leviedOn = "Cars";
+        t4.rate = 0.025f;
+        tariffs.save(t4);
+
+        Tariff t5 = new Tariff();
+        t5.leviedAgainst = "Canada";
+        t5.leviedBy = "USA";
+        t5.leviedOn = "Lumber";
+        t5.rate = 0.1799f;
+        tariffs.save(t5);
+
+        Tariff t6 = new Tariff();
+        t6.leviedAgainst = "Bangladesh";
+        t6.leviedBy = "USA";
+        t6.leviedOn = "Textiles";
+        t6.rate = 0.158f;
+        tariffs.save(t6);
+
+        Tariff t7 = new Tariff();
+        t7.leviedAgainst = "Mexico";
+        t7.leviedBy = "USA";
+        t7.leviedOn = "Trucks";
+        t7.rate = 0.25f;
+        tariffs.save(t7);
+
+        Tariff t8 = new Tariff();
+        t8.leviedAgainst = "Canada";
+        t8.leviedBy = "USA";
+        t8.leviedOn = "Copper";
+        t8.rate = 0.0194f;
+        tariffs.save(t8);
+
+        assertIterableEquals(List.of("Copper", "Lumber"),
+                             tariffs.findByLeviedAgainst("Canada").map(o -> o.leviedOn).sorted().collect(Collectors.toList()));
+
+        // Iterator with offset pagination:
+        Iterator<Tariff> it = tariffs.findByLeviedAgainstLessThanOrderByKeyDesc("M", Pageable.ofSize(3));
+
+        Tariff t;
+        assertEquals(true, it.hasNext());
+        assertNotNull(t = it.next());
+        assertEquals(t8.leviedAgainst, t.leviedAgainst);
+        Long t8key = t.key;
+
+        assertEquals(true, it.hasNext());
+        assertNotNull(t = it.next());
+        assertEquals(t6.leviedAgainst, t.leviedAgainst);
+
+        assertEquals(true, it.hasNext());
+        assertNotNull(t = it.next());
+        assertEquals(t5.leviedAgainst, t.leviedAgainst);
+
+        assertEquals(true, it.hasNext());
+        assertEquals(true, it.hasNext());
+        assertNotNull(t = it.next());
+        assertEquals(t4.leviedAgainst, t.leviedAgainst);
+
+        assertNotNull(t = it.next());
+        assertEquals(t3.leviedAgainst, t.leviedAgainst);
+
+        assertNotNull(t = it.next());
+        assertEquals(t2.leviedAgainst, t.leviedAgainst);
+        Long t2key = t.key;
+
+        assertNotNull(t = it.next());
+        assertEquals(t1.leviedAgainst, t.leviedAgainst);
+
+        assertEquals(false, it.hasNext());
+        assertEquals(false, it.hasNext());
+
+        // Iterator with keyset pagination:
+        it = tariffs.findByLeviedAgainstLessThanOrderByKeyDesc("M", Pageable.ofSize(2).afterKeyset(t8key));
+
+        assertNotNull(t = it.next());
+        assertEquals(t6.leviedAgainst, t.leviedAgainst);
+
+        assertNotNull(t = it.next());
+        assertEquals(t5.leviedAgainst, t.leviedAgainst);
+
+        assertNotNull(t = it.next());
+        assertEquals(t4.leviedAgainst, t.leviedAgainst);
+
+        assertEquals(true, it.hasNext());
+        assertNotNull(t = it.next());
+        assertEquals(t3.leviedAgainst, t.leviedAgainst);
+
+        assertNotNull(t = it.next());
+        assertEquals(t2.leviedAgainst, t.leviedAgainst);
+
+        assertEquals(true, it.hasNext());
+        assertNotNull(t = it.next());
+        assertEquals(t1.leviedAgainst, t.leviedAgainst);
+
+        assertEquals(false, it.hasNext());
+
+        // Iterator with keyset pagination obtaining pages in reverse direction
+        it = tariffs.findByLeviedAgainstLessThanOrderByKeyDesc("M", Pageable.ofSize(2).beforeKeyset(t2key));
+
+        assertEquals(true, it.hasNext());
+        assertNotNull(t = it.next());
+        assertEquals(t4.leviedAgainst, t.leviedAgainst);
+
+        assertNotNull(t = it.next());
+        assertEquals(t3.leviedAgainst, t.leviedAgainst);
+
+        assertNotNull(t = it.next());
+        assertEquals(t6.leviedAgainst, t.leviedAgainst);
+
+        assertEquals(true, it.hasNext());
+        assertNotNull(t = it.next());
+        assertEquals(t5.leviedAgainst, t.leviedAgainst);
+
+        assertNotNull(t = it.next());
+        assertEquals(t8.leviedAgainst, t.leviedAgainst);
+
+        assertEquals(false, it.hasNext());
+
+        // Paginated iterator with no results:
+        it = tariffs.findByLeviedAgainstLessThanOrderByKeyDesc("A", Pageable.ofSize(3));
+        assertEquals(false, it.hasNext());
+
+        t = tariffs.findByLeviedByAndLeviedAgainstAndLeviedOn("USA", "Bangladesh", "Textiles");
+        assertEquals(t6.rate, t.rate, 0.0001f);
+
+        // List return type for Pagination only represents a single page, not all pages.
+        // page 1:
+        assertIterableEquals(List.of("China", "Germany", "India", "Japan"),
+                             tariffs.findByLeviedByOrderByKey("USA", Pageable.ofSize(4))
+                                             .stream()
+                                             .map(o -> o.leviedAgainst)
+                                             .collect(Collectors.toList()));
+        // page 2:
+        assertIterableEquals(List.of("Canada", "Bangladesh", "Mexico", "Canada"),
+                             tariffs.findByLeviedByOrderByKey("USA", Pageable.ofSize(4).page(2))
+                                             .stream()
+                                             .map(o -> o.leviedAgainst)
+                                             .collect(Collectors.toList()));
+
+        // Random access to paginated list:
+        List<Tariff> list = tariffs.findByLeviedByOrderByKey("USA", Pageable.ofPage(1));
+        assertEquals(t4.leviedAgainst, list.get(3).leviedAgainst);
+        assertEquals(t7.leviedAgainst, list.get(6).leviedAgainst);
+        assertEquals(t2.leviedAgainst, list.get(1).leviedAgainst);
+        assertEquals(t8.leviedAgainst, list.get(7).leviedAgainst);
+
+        assertEquals(8, tariffs.deleteByLeviedBy("USA"));
     }
 
     /**
